refactor: extrait buildJobsMinQuerystring partagé client/serveur (#5033)
